requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
langchain>=0.0.350
langchain-community>=0.0.10
langchain-google-genai>=1.0.0
faiss-cpu>=1.7.4
plotly>=5.17.0
pandas>=2.0.0
numpy>=1.24.0
exa-py>=1.0.0
sentence-transformers>=2.2.0
chromadb>=0.4.0